Triple values can now be nil.
When a Triplestore is asked for a triple for a entity/attribute combination, and none exists, it creates a new triple with value nil.
If the value of that triple is updated, it will be saved in the persistent store.
If the value of an triple is set to nil, it will be deleted from the pesistent store.

So, the persistent store still doesn't store nil-valued triples. But the Triplestore may create and index such triple objects.